+++ src/lib/pages/block-details/components/BlockDetailsTop.tsx
@@ -7,7 +7,8 @@ import {
   Button,
 } from "@chakra-ui/react";
 
-import { useBaseApiRoute, useInternalNavigate } from "lib/app-provider";
+import { useBaseApiRoute } from "lib/app-provider";
+import { AppLink } from "lib/components/AppLink";
 import { CopyLink } from "lib/components/CopyLink";
 import { DotSeparator } from "lib/components/DotSeparator";
 import { CustomIcon } from "lib/components/icon";
@@ -29,7 +30,7 @@ interface BlockDetailsTopProps {
 export const BlockDetailsTop = ({ blockData }: BlockDetailsTopProps) => {
   const block = Number(blockData.height);
   const lcdEndpoint = useBaseApiRoute("rest");
-  const navigate = useInternalNavigate();
+
   const openLcdPage = () =>
     openNewTab(
       `${lcdEndpoint}/cosmos/base/tendermint/v1beta1/blocks/${blockData.height}`
@@ -37,78 +38,84 @@ export const BlockDetailsTop = ({ blockData }: BlockDetailsTopProps) => {
   const disablePrevious = block <= 1;
   return (
     <Flex
-      mb={12}
-      pb={12}
-      mt={6}
+      justify="space-between"
+      mb={{ base: 8, md: 12 }}
+      pb={{ base: 8, md: 12 }}
       borderBottomColor="gray.700"
       borderBottomWidth="1px"
     >
-      <Flex direction="column" gap={2} width="full">
-        <Flex justifyContent="space-between" alignItems="center" width="full">
-          <Flex alignItems="center">
+      <Flex direction="column" gap={1} w="full">
+        <Flex
+          justify="space-between"
+          align="center"
+          width="full"
+          mt={{ base: 2, md: 5 }}
+          mb={{ base: 2, md: 0 }}
+        >
+          <Flex align="center">
             <CustomIcon name="block" boxSize={5} color="secondary.main" />
-            <Heading as="h5" variant="h5" className="ellipsis">
+            <Heading
+              as="h5"
+              variant={{ base: "h6", md: "h5" }}
+              className="ellipsis"
+            >
               {blockData.height}
             </Heading>
           </Flex>
-          <Flex gap={2}>
+          <Flex gap={{ base: 1, md: 2 }} align="center">
             {!disablePrevious && (
-              <StyledIconButton
-                icon={<CustomIcon name="chevron-left" />}
-                variant="ghost-gray"
-                onClick={() =>
-                  navigate({
-                    pathname: "/blocks/[height]",
-                    query: { height: block - 1 },
-                  })
-                }
-              />
+              <AppLink href={`/blocks/${block - 1}`}>
+                <StyledIconButton
+                  icon={<CustomIcon name="chevron-left" />}
+                  variant="ghost-gray"
+                />
+              </AppLink>
             )}
-            <StyledIconButton
-              icon={<CustomIcon name="chevron-right" />}
-              variant="ghost-gray"
-              onClick={() =>
-                navigate({
-                  pathname: "/blocks/[height]",
-                  query: { height: block + 1 },
-                })
-              }
-            />
+            <AppLink href={`/blocks/${block + 1}`}>
+              <StyledIconButton
+                icon={<CustomIcon name="chevron-right" />}
+                variant="ghost-gray"
+              />
+            </AppLink>
             <Button
               variant="ghost-gray"
-              padding={2}
-              rightIcon={
-                <CustomIcon name="launch" boxSize={3} color="text.dark" />
-              }
+              size={{ base: "sm", md: "md" }}
+              rightIcon={<CustomIcon name="launch" boxSize={3} />}
               onClick={openLcdPage}
             >
               View in JSON
             </Button>
           </Flex>
         </Flex>
-        <Flex direction="column" gap={1}>
-          <Flex gap={2}>
-            <Text variant="body2" color="text.dark">
-              Block Hash:
+        <Flex
+          gap={{ base: 0, md: 2 }}
+          direction={{ base: "column", md: "row" }}
+        >
+          <Text
+            variant="body2"
+            fontWeight={500}
+            color="text.dark"
+            display="inline"
+          >
+            Block Hash:
+          </Text>
+          <CopyLink
+            value={blockData.hash.toUpperCase()}
+            amptrackSection="block_details_top"
+            type="block_hash"
+          />
+        </Flex>
+        <Flex gap={2} mt={1} alignItems="center">
+          <Flex gap={1} alignItems="center">
+            <CustomIcon name="history" boxSize={3} color="gray.600" />
+            <Text variant={{ base: "body3", md: "body2" }} color="text.dark">
+              {dateFromNow(blockData.timestamp)}
             </Text>
-            <CopyLink
-              value={blockData.hash.toUpperCase()}
-              amptrackSection="block_details_top"
-              type="block_hash"
-            />
           </Flex>
-          <Flex gap={2} alignItems="center">
-            <Flex gap={1} alignItems="center">
-              <CustomIcon name="history" boxSize={3} color="gray.600" />
-              <Text variant="body2" color="text.dark">
-                {dateFromNow(blockData.timestamp)}
-              </Text>
-            </Flex>
-            <DotSeparator />
-            <Text variant="body2" color="text.dark">
-              {formatUTC(blockData.timestamp)}
-            </Text>
-          </Flex>
+          <DotSeparator />
+          <Text variant={{ base: "body3", md: "body2" }} color="text.dark">
+            {formatUTC(blockData.timestamp)}
+          </Text>
         </Flex>
       </Flex>
     </Flex>