+++ src/lib/pages/public-project/slug.tsx
@@ -54,7 +54,12 @@ export const ProjectDetail = () => {
     <PageContainer>
       <DetailHeader details={projectDetail} slug={slug} />
       <Tabs index={tabIndex} isLazy lazyBehavior="keepMounted">
-        <TabList my={6} borderBottom="1px" borderColor="gray.800">
+        <TabList
+          my={6}
+          borderBottom="1px solid"
+          borderColor="gray.700"
+          overflowX="scroll"
+        >
           <CustomTab
             count={overviewCount}
             onClick={() => setTabIndex(TabIndex.Overview)}
@@ -85,7 +90,6 @@ export const ProjectDetail = () => {
             Accounts
           </CustomTab>
         </TabList>
-
         <TabPanels my={8}>
           <TabPanel p={0}>
             {wasm.enabled && (